move cfn tag logic to new file
